requirements.txt
Django>=4.2
gunicorn
psycopg2-binary
dj-database-url
python-dotenv
whitenoise
Pillow
PyPDF2
requests
reportlab